+++ tests/testthat/helper-utils.R
@@ -88,7 +88,7 @@ define_and_compile_model <- function() {
   model
 }
 
-random_array <- function(dim) {
+random_array <- function(..., dim = unlist(c(...))) {
   array(runif(prod(dim)), dim = dim)
 }